del初始被设定为了159.94，不知道是在哪

+++ stm32f103c8t6/Core/Src/Coordinate.c
@@ -50,12 +50,12 @@ SystemParams sys_set = {
 void Coordinate_Init(void)
 {
     //比例计算
-    sys_set.x_pixel_to_mm = (
+    sys_set.x_pixel_to_mm = (float)(
         (sys_set.Calib_Point[0].x - sys_set.Calib_Point[1].x) + 
         (sys_set.Calib_Point[3].x - sys_set.Calib_Point[2].x)
     ) / 2 / 500;
 
-    sys_set.y_pixel_to_mm = (
+    sys_set.y_pixel_to_mm = (float)(
         (sys_set.Calib_Point[3].y - sys_set.Calib_Point[0].y) + 
         (sys_set.Calib_Point[2].y - sys_set.Calib_Point[1].y)
     ) / 2 / 500;
